fix: change ready behaviour
Improve code on event ready with code splitting
